pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jobhunter"
version = "0.1.0"
description = "Automated job search and application system"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "playwright>=1.49.0",
    "patchright>=1.0.0",
    "playwright-stealth>=2.0.0",
    "cryptography>=43.0.0",
    "apscheduler>=3.10.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0.0",
    "google-api-python-client>=2.155.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.0.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[project.scripts]
jobhunter = "jobhunter.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/jobhunter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.hatch.metadata]
allow-direct-references = true
